+++ src/mastodon_api/subscriptions/views.rs
@@ -17,6 +17,7 @@ use mitra_models::{
     database::{get_database_client, DbPool},
     invoices::queries::{
         create_invoice,
+        create_remote_invoice,
         get_invoice_by_id,
         set_invoice_status,
     },
@@ -30,6 +31,7 @@ use mitra_models::{
         PaymentOption,
         PaymentType,
         ProfileUpdateData,
+        RemoteMoneroSubscription,
     },
     subscriptions::queries::get_subscription_by_participants,
     users::queries::get_user_by_id,
@@ -220,34 +222,54 @@ async fn create_invoice_view(
     db_pool: web::Data<DbPool>,
     invoice_data: web::Json<InvoiceData>,
 ) -> Result<HttpResponse, MastodonError> {
-    let monero_config = config.monero_config()
-        .ok_or(MastodonError::NotSupported)?;
     if invoice_data.sender_id == invoice_data.recipient_id {
         return Err(ValidationError("sender must be different from recipient").into());
-    };
-    if invoice_data.chain_id != monero_config.chain_id {
-        return Err(ValidationError("unexpected chain ID").into());
     };
     validate_amount(invoice_data.amount)?;
     let db_client = &**get_database_client(&db_pool).await?;
     let sender = get_profile_by_id(db_client, &invoice_data.sender_id).await?;
-    let recipient = get_user_by_id(db_client, &invoice_data.recipient_id).await?;
-    let _subscription_option: MoneroSubscription = recipient.profile
-        .payment_options
-        .find_subscription_option(&invoice_data.chain_id)
-        .ok_or(ValidationError("recipient can't accept payment"))?;
+    let recipient = get_profile_by_id(db_client, &invoice_data.recipient_id).await?;
 
-    let payment_address = create_monero_address(monero_config).await
-        .map_err(|_| MastodonError::InternalError)?
-        .to_string();
-    let db_invoice = create_invoice(
-        db_client,
-        &sender.id,
-        &recipient.id,
-        &invoice_data.chain_id,
-        &payment_address,
-        invoice_data.amount,
-    ).await?;
+    let db_invoice = if recipient.is_local() {
+        // Local recipient
+        let monero_config = config.monero_config()
+            .ok_or(MastodonError::NotSupported)?;
+        if invoice_data.chain_id != monero_config.chain_id {
+            return Err(ValidationError("unexpected chain ID").into());
+        };
+        let _subscription_option: MoneroSubscription = recipient
+            .payment_options
+            .find_subscription_option(&invoice_data.chain_id)
+            .ok_or(ValidationError("recipient can't accept payment"))?;
+        let payment_address = create_monero_address(monero_config).await
+            .map_err(|_| MastodonError::InternalError)?
+            .to_string();
+        create_invoice(
+            db_client,
+            &sender.id,
+            &recipient.id,
+            &invoice_data.chain_id,
+            &payment_address,
+            invoice_data.amount,
+        ).await?
+    } else {
+        // Remote recipient; the sender must be local
+        let sender = get_user_by_id(db_client, &sender.id).await?;
+        let subscription_option: RemoteMoneroSubscription = recipient
+            .payment_options
+            .find_subscription_option(&invoice_data.chain_id)
+            .ok_or(ValidationError("recipient can't accept payment"))?;
+        if !subscription_option.fep_0837_enabled {
+            return Err(MastodonError::OperationError("recipient doesn't support FEP-0837"));
+        };
+        create_remote_invoice(
+            db_client,
+            &sender.id,
+            &recipient.id,
+            &invoice_data.chain_id,
+            invoice_data.amount,
+        ).await?
+    };
     let invoice = Invoice::from(db_invoice);
     Ok(HttpResponse::Ok().json(invoice))
 }